Clean up and type fixes.
After discussions with Keith Bostic I've learned that always specifying the
full type (u_int32_t rather than int) as a general rule is not always a good
idea.  Now, as I review code I'm reverting some of those changes.  The new
rule of thumb is to use explict typing only in places that require it.